+++ bindings/python/FloatingBaseEstimators/src/LeggedOdometry.cpp
@@ -13,6 +13,9 @@
 #include <BipedalLocomotion/FloatingBaseEstimators/LeggedOdometry.h>
 
 #include <BipedalLocomotion/bindings/FloatingBaseEstimators/LeggedOdometry.h>
+#include <BipedalLocomotion/bindings/type_caster/swig.h>
+
+#include <iDynTree/KinDynComputations.h>
 
 namespace BipedalLocomotion
 {
@@ -37,8 +40,18 @@ void CreateLeggedOdometry(pybind11::module& module)
             "initialize",
             [](LeggedOdometry& impl,
                std::shared_ptr<IParametersHandler> handler,
-               std::shared_ptr<iDynTree::KinDynComputations> kinDyn) -> bool {
-                return impl.initialize(handler, kinDyn);
+               py::object& obj) -> bool {
+                std::shared_ptr<iDynTree::KinDynComputations>* kinDynPtr
+                    = pybind11::detail::swig_wrapped_pointer_to_pybind<
+                        std::shared_ptr<iDynTree::KinDynComputations>>(obj);
+
+                if (kinDynPtr == nullptr)
+                {
+                    throw ::pybind11::value_error("Invalid input for the function. Please provide "
+                                                  "an iDynTree::KinDynComputations object.");
+                }
+
+                return impl.initialize(handler, *kinDynPtr);
             },
             py::arg("handler"),
             py::arg("kindyn"))