+++ bindings/python/TSID/src/SO3Task.cpp
@@ -9,9 +9,12 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 
-#include <BipedalLocomotion/TSID/TSIDLinearTask.h>
 #include <BipedalLocomotion/TSID/SO3Task.h>
+#include <BipedalLocomotion/TSID/TSIDLinearTask.h>
+
+#include <BipedalLocomotion/bindings/System/LinearTask.h>
 #include <BipedalLocomotion/bindings/TSID/SO3Task.h>
+#include <BipedalLocomotion/bindings/TSID/TSIDLinearTask.h>
 
 namespace BipedalLocomotion
 {
@@ -27,7 +30,9 @@ void CreateSO3Task(pybind11::module& module)
 
     py::class_<SO3Task, std::shared_ptr<SO3Task>, TSIDLinearTask>(module, "SO3Task")
         .def(py::init())
-        .def("set_kin_dyn", &SO3Task::setKinDyn, py::arg("kin_dyn"))
+        .def("set_kin_dyn",
+             BipedalLocomotion::bindings::System::setKinDyn<SO3Task>,
+             py::arg("kin_dyn"))
         .def("set_set_point",
              &SO3Task::setSetPoint,
              py::arg("I_R_F"),